requirements.txt
SQLAlchemy==2.0.19
streamlit==1.12.0
streamlit-chat==0.1.1
langchain==0.0.312
boto3==1.28.61
botocore==1.31.61
altair==4.0.0
pydantic==1.10.9
psycopg==3.1.10
psycopg-binary==3.1.11
psycopg2-binary==2.9.6
pgvector==0.2.3
fake-useragent==1.3.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
youtube-transcript-api
unstructured
python-pptx
unstructured[pdf]
python-dotenv==1.0.0
